requirements.txt
streamlit
google-genai
python-dotenv
numpy
scikit-learn
pytest
sentence-transformers